refactor: allow lbc/category template to have multiple page layouts
ACK-227
