+++ src/pages/Lend/LoanChannelCategoryExperiment.vue
@@ -1,0 +1,5 @@
+<template>
+	<div>
+		<!-- This template will be the page layout for ACK-227 -->
+	</div>
+</template>
